+++ internal/database/ent/db/account/account.go
@@ -93,11 +93,11 @@ func ValidColumn(column string) bool {
 
 var (
 	// DefaultCreatedAt holds the default value on creation for the "created_at" field.
-	DefaultCreatedAt func() int
+	DefaultCreatedAt func() int64
 	// DefaultUpdatedAt holds the default value on creation for the "updated_at" field.
-	DefaultUpdatedAt func() int
+	DefaultUpdatedAt func() int64
 	// UpdateDefaultUpdatedAt holds the default value on update for the "updated_at" field.
-	UpdateDefaultUpdatedAt func() int
+	UpdateDefaultUpdatedAt func() int64
 	// RealmValidator is a validator for the "realm" field. It is called by the builders before save.
 	RealmValidator func(string) error
 	// NicknameValidator is a validator for the "nickname" field. It is called by the builders before save.
