+++ internal/database/ent/db/accountsnapshot/where.go
@@ -65,17 +65,17 @@ func IDContainsFold(id string) predicate.AccountSnapshot {
 }
 
 // CreatedAt applies equality check predicate on the "created_at" field. It's identical to CreatedAtEQ.
-func CreatedAt(v int) predicate.AccountSnapshot {
+func CreatedAt(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldEQ(FieldCreatedAt, v))
 }
 
 // UpdatedAt applies equality check predicate on the "updated_at" field. It's identical to UpdatedAtEQ.
-func UpdatedAt(v int) predicate.AccountSnapshot {
+func UpdatedAt(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldEQ(FieldUpdatedAt, v))
 }
 
 // LastBattleTime applies equality check predicate on the "last_battle_time" field. It's identical to LastBattleTimeEQ.
-func LastBattleTime(v int) predicate.AccountSnapshot {
+func LastBattleTime(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldEQ(FieldLastBattleTime, v))
 }
 
@@ -100,82 +100,82 @@ func RegularBattles(v int) predicate.AccountSnapshot {
 }
 
 // CreatedAtEQ applies the EQ predicate on the "created_at" field.
-func CreatedAtEQ(v int) predicate.AccountSnapshot {
+func CreatedAtEQ(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldEQ(FieldCreatedAt, v))
 }
 
 // CreatedAtNEQ applies the NEQ predicate on the "created_at" field.
-func CreatedAtNEQ(v int) predicate.AccountSnapshot {
+func CreatedAtNEQ(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldNEQ(FieldCreatedAt, v))
 }
 
 // CreatedAtIn applies the In predicate on the "created_at" field.
-func CreatedAtIn(vs ...int) predicate.AccountSnapshot {
+func CreatedAtIn(vs ...int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldIn(FieldCreatedAt, vs...))
 }
 
 // CreatedAtNotIn applies the NotIn predicate on the "created_at" field.
-func CreatedAtNotIn(vs ...int) predicate.AccountSnapshot {
+func CreatedAtNotIn(vs ...int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldNotIn(FieldCreatedAt, vs...))
 }
 
 // CreatedAtGT applies the GT predicate on the "created_at" field.
-func CreatedAtGT(v int) predicate.AccountSnapshot {
+func CreatedAtGT(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldGT(FieldCreatedAt, v))
 }
 
 // CreatedAtGTE applies the GTE predicate on the "created_at" field.
-func CreatedAtGTE(v int) predicate.AccountSnapshot {
+func CreatedAtGTE(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldGTE(FieldCreatedAt, v))
 }
 
 // CreatedAtLT applies the LT predicate on the "created_at" field.
-func CreatedAtLT(v int) predicate.AccountSnapshot {
+func CreatedAtLT(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldLT(FieldCreatedAt, v))
 }
 
 // CreatedAtLTE applies the LTE predicate on the "created_at" field.
-func CreatedAtLTE(v int) predicate.AccountSnapshot {
+func CreatedAtLTE(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldLTE(FieldCreatedAt, v))
 }
 
 // UpdatedAtEQ applies the EQ predicate on the "updated_at" field.
-func UpdatedAtEQ(v int) predicate.AccountSnapshot {
+func UpdatedAtEQ(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldEQ(FieldUpdatedAt, v))
 }
 
 // UpdatedAtNEQ applies the NEQ predicate on the "updated_at" field.
-func UpdatedAtNEQ(v int) predicate.AccountSnapshot {
+func UpdatedAtNEQ(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldNEQ(FieldUpdatedAt, v))
 }
 
 // UpdatedAtIn applies the In predicate on the "updated_at" field.
-func UpdatedAtIn(vs ...int) predicate.AccountSnapshot {
+func UpdatedAtIn(vs ...int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldIn(FieldUpdatedAt, vs...))
 }
 
 // UpdatedAtNotIn applies the NotIn predicate on the "updated_at" field.
-func UpdatedAtNotIn(vs ...int) predicate.AccountSnapshot {
+func UpdatedAtNotIn(vs ...int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldNotIn(FieldUpdatedAt, vs...))
 }
 
 // UpdatedAtGT applies the GT predicate on the "updated_at" field.
-func UpdatedAtGT(v int) predicate.AccountSnapshot {
+func UpdatedAtGT(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldGT(FieldUpdatedAt, v))
 }
 
 // UpdatedAtGTE applies the GTE predicate on the "updated_at" field.
-func UpdatedAtGTE(v int) predicate.AccountSnapshot {
+func UpdatedAtGTE(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldGTE(FieldUpdatedAt, v))
 }
 
 // UpdatedAtLT applies the LT predicate on the "updated_at" field.
-func UpdatedAtLT(v int) predicate.AccountSnapshot {
+func UpdatedAtLT(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldLT(FieldUpdatedAt, v))
 }
 
 // UpdatedAtLTE applies the LTE predicate on the "updated_at" field.
-func UpdatedAtLTE(v int) predicate.AccountSnapshot {
+func UpdatedAtLTE(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldLTE(FieldUpdatedAt, v))
 }
 
@@ -210,42 +210,42 @@ func TypeNotIn(vs ...models.SnapshotType) predicate.AccountSnapshot {
 }
 
 // LastBattleTimeEQ applies the EQ predicate on the "last_battle_time" field.
-func LastBattleTimeEQ(v int) predicate.AccountSnapshot {
+func LastBattleTimeEQ(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldEQ(FieldLastBattleTime, v))
 }
 
 // LastBattleTimeNEQ applies the NEQ predicate on the "last_battle_time" field.
-func LastBattleTimeNEQ(v int) predicate.AccountSnapshot {
+func LastBattleTimeNEQ(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldNEQ(FieldLastBattleTime, v))
 }
 
 // LastBattleTimeIn applies the In predicate on the "last_battle_time" field.
-func LastBattleTimeIn(vs ...int) predicate.AccountSnapshot {
+func LastBattleTimeIn(vs ...int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldIn(FieldLastBattleTime, vs...))
 }
 
 // LastBattleTimeNotIn applies the NotIn predicate on the "last_battle_time" field.
-func LastBattleTimeNotIn(vs ...int) predicate.AccountSnapshot {
+func LastBattleTimeNotIn(vs ...int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldNotIn(FieldLastBattleTime, vs...))
 }
 
 // LastBattleTimeGT applies the GT predicate on the "last_battle_time" field.
-func LastBattleTimeGT(v int) predicate.AccountSnapshot {
+func LastBattleTimeGT(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldGT(FieldLastBattleTime, v))
 }
 
 // LastBattleTimeGTE applies the GTE predicate on the "last_battle_time" field.
-func LastBattleTimeGTE(v int) predicate.AccountSnapshot {
+func LastBattleTimeGTE(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldGTE(FieldLastBattleTime, v))
 }
 
 // LastBattleTimeLT applies the LT predicate on the "last_battle_time" field.
-func LastBattleTimeLT(v int) predicate.AccountSnapshot {
+func LastBattleTimeLT(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldLT(FieldLastBattleTime, v))
 }
 
 // LastBattleTimeLTE applies the LTE predicate on the "last_battle_time" field.
-func LastBattleTimeLTE(v int) predicate.AccountSnapshot {
+func LastBattleTimeLTE(v int64) predicate.AccountSnapshot {
 	return predicate.AccountSnapshot(sql.FieldLTE(FieldLastBattleTime, v))
 }
